requirements.txt
Flask
python-dotenv
requests
sqlite3